2 staff user denemeleri

+++ Admin/views.py
@@ -73,10 +73,14 @@ def indexAdmin(request):
 def loginAdmin(request):
     if request.method == "POST":
         form = AuthenticationForm(request, data=request.POST)		
+        print("validin üstünde")
+        print(request.POST)
         if form.is_valid():
+            print("valid kısmında")
             username = form.cleaned_data.get('username')
             password = form.cleaned_data.get('password')
             user = authenticate(username=username, password=password)
+            print(user.is_staff)
             if user is not None and user.is_staff:
                 login(request, user)
                 return redirect("indexAdmin")
@@ -998,7 +1002,7 @@ def gizlilikMetniAdmin(request):
 
 @permission_required('is_staff',login_url="loginAdmin")
 def blogListAdmin(request):
-    blogs=blogModel.objects.all().order_by("created_date")
+    blogs=blogModel.objects.filter(author=request.user).order_by("created_date")
     context={
         "blogs":blogs,
     }
@@ -1285,9 +1289,11 @@ def pastAppointmentsAdmin(request):
             if i.finishing_time>datetime.datetime.now().time():
                 pass
             else:
-                appointments.append(i)
+                if i.top.whois==request.user:
+                    appointments.append(i)
         else:
-            appointments.append(i)
+            if i.top.whois==request.user:
+                appointments.append(i)
    
     context={
         "appointments":appointments,
@@ -1312,9 +1318,11 @@ def nextAppointmentsAdmin(request):
             if i.finishing_time<datetime.datetime.now().time():
                 pass
             else:
-                appointments.append(i)
+                if i.top.whois==request.user:
+                    appointments.append(i)
         else:
-            appointments.append(i)
+            if i.top.whois==request.user:
+                    appointments.append(i)
    
     context={
         "appointments":appointments,
@@ -1329,7 +1337,11 @@ def nextAppointmentsAdmin(request):
 
 @permission_required('is_staff',login_url="loginAdmin")
 def listAppointmentsAdmin(request):
-    appointments=appointmentModel.objects.filter(status="pending").order_by("-date")
+    appointments1=appointmentModel.objects.filter(status="pending").order_by("-date")
+    appointments=list()
+    for i in appointments1:
+        if i.top.whois==request.user:
+            appointments.append(i)
     context={
         "appointments":appointments,
         "type":"requests",
@@ -1436,13 +1448,13 @@ def showVideoStatusDetail(request,slug):
 
 @permission_required('is_staff',login_url="loginAdmin")
 def appointmentsAdmin(request):
-    appointments=appointmentAdminModel.objects.all().order_by("date")
+    appointments=appointmentAdminModel.objects.filter(whois=request.user).order_by("date")
     for i in appointments:
         if i.date<datetime.datetime.today().date():
             for j in i.bottomsAppo.all():
                 if j.status=="no":
                     j.delete()
-    appointments=appointmentAdminModel.objects.all().order_by("date")
+    appointments=appointmentAdminModel.objects.filter(whois=request.user).order_by("date")
     context={
         "appointments":appointments,
     }
@@ -1458,7 +1470,8 @@ def addAppointmentsAdmin(request):
     if request.method == "POST":   
         form = appointmentAdminModelForm(request.POST, request.FILES or None)		
         if form.is_valid(): 
-            data=form.save(commit=False)         
+            data=form.save(commit=False)   
+            data.whois=request.user      
             finish=form.cleaned_data.get('finishing_time')
             start=form.cleaned_data.get('starting_time')
             date=form.cleaned_data.get('date')
@@ -1474,7 +1487,7 @@ def addAppointmentsAdmin(request):
             difference=datetime.datetime.combine(date.today(), finish.time()) - datetime.datetime.combine(date.today(), start.time())
             dd=str(difference).split(":")
             x=0
-            oldAppointments=appointmentAdminModel.objects.filter(date=date)
+            oldAppointments=appointmentAdminModel.objects.filter(date=date,whois=request.user)
             oldAppointmentListTimes=list()
             if oldAppointments:
                 oldAppoinment=oldAppointments.first()
@@ -1512,8 +1525,11 @@ def addAppointmentsAdmin(request):
 @permission_required('is_staff',login_url="loginAdmin")
 def deleteAppointmentOfAdmin(request,pk):
     obj=get_object_or_404(appointmentAdminModel,pk=pk)
-    obj.delete()
-    messages.success(request,"Randevu tarihi ve o tarihli randevular başarıyla silindi")
+    if obj.whois==request.user:
+        obj.delete()
+        messages.success(request,"Randevu tarihi ve o tarihli randevular başarıyla silindi")
+    else:
+        messages.error(request,"Randevu size ait değildir.Silme yetkiniz bulunmamaktadır.")
     return redirect(request.META['HTTP_REFERER']) 
 
 
@@ -1583,7 +1599,7 @@ def editAppointmentsScheduleAdmin(request,pk):
             if date<datetime.datetime.today().date():
                 messages.error(request,"Girmiş olduğunuz tarih bugünden önce olamaz")
                 return redirect("showAppointmentsScheduleAdmin",apt.top.pk)
-            appointmentAdminModel.objects.create(date=date,starting_time=t.time(),finishing_time=(t+datetime.timedelta(hours=1)).time())
+            appointmentAdminModel.objects.create(whois=request.user,date=date,starting_time=t.time(),finishing_time=(t+datetime.timedelta(hours=1)).time())
         if isSame=="false":
             apt.starting_time=t.time()
             apt.finishing_time=(t+datetime.timedelta(hours=1)).time()
@@ -1620,7 +1636,7 @@ def showDetailOrderAdmin(request,pk):
 
 @permission_required('is_staff',login_url="loginAdmin")
 def appointmentCategoryListAdmin(request):
-    categories=appointmentCategoryModel.objects.all().order_by("created_date")
+    categories=appointmentCategoryModel.objects.filter(whois=request.user).order_by("created_date")
     context={
         "categories":categories,
     }
@@ -1642,7 +1658,9 @@ def addAppointmentCategoryAdmin(request):
         else:
             form = appointmentCategoryModelForm(request.POST, request.FILES or None)	
         if form.is_valid(): 
-            form.save()            
+            dd=form.save(commit=False)      
+            dd.whois=request.user
+            dd.save()      
             messages.success(request,"Randevu kategorisi başarıyla kaydedildi.")
             return redirect("appointmentCategoryListAdmin")
         else:
